Add dedicated columns for recipe scores
Began as an attempt to reduce excessive database calls and time spent
processing the scores for the recipe card progress bars. Originally
the scores were calculated off of cook_time + prep_time, for example,
but this was called every time a search query page was loaded, for
each card on the page. Thus, this commit, which adds separate columns
to the Recipe table for holding the relevant scores (specifically
time and ingredients, as difficulty was already just 1 number), since
they really only need to be calculated whenever the Recipe is created
or updated. Now that this is in place, a single database call can pull
all the scores for a query at once, though that is left for a future
commit.